+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add new functionality to filters (and, or, brackets).
 - Add shorter aliases for sorting and filtering.
 - Add option to sort songs as part of query
+- Add terminal friendly alternatives for descending order.
 
 ## v0.5.4
 - Add option to enable/disable color in terminal.